@@ -1,6 +1,6 @@
 {
   "name": "shop-dev",
-  "version": "0.1.16",
+  "version": "0.1.18",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -14,16 +14,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/schematics": "^17.3.11",
-    "@angular/animations": "^17.3.12",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/compiler-cli": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
+    "@angular-devkit/schematics": "^18.2.14",
+    "@angular/animations": "^18.2.13",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/compiler-cli": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/router": "^18.2.13",
     "@ngtools/webpack": "^16.2.0",
     "@ngx-translate/core": "^15.0.0",
     "@ngx-translate/http-loader": "^3.0.1",
@@ -38,10 +38,10 @@
     "zone.js": "^0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.11",
-    "@angular/cli": "^17.3.11",
-    "@angular/language-service": "^17.3.12",
-    "@angular/localize": "^17.3.12",
+    "@angular-devkit/build-angular": "^18.2.14",
+    "@angular/cli": "^18.2.14",
+    "@angular/language-service": "^18.2.13",
+    "@angular/localize": "^18.2.13",
     "@cypress/schematic": "^1.7.0",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.10",
